% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knapsack_FUNS.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Algorithm for the 0-1 knapsack problem}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{a dataframe with w (weights) as first column and v (values) as second column.}

\item{W}{a scalar representing the limiting capacity of the knapsack}
}
\value{
a list of the maximum value and the elements used to obtail this value
}
\description{
Algorithm for the 0-1 knapsack problem
}
\examples{
set.seed (42)
n <- 2000
knapsack_objects <- data.frame(w = sample (1: 4000, size = n, replace = TRUE), v = runif(n = n, 0, 10000))
brute_force_knapsack(x = knapsack_objects[1:8, ], W = 3500)
}
