% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knapsack_FUNS.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy aproximation algorithm for the knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{a dataframe with w (weights) as first column and v (values) as second column.}

\item{W}{a scalar representing the limiting capacity of the knapsack}
}
\value{
a list of the maximum value and the elements used to obtail this value
}
\description{
Greedy aproximation algorithm for the knapsack problem
}
\examples{
set.seed (42)
n <- 2000
knapsack_objects <- data.frame(w = sample (1: 4000, size = n, replace = TRUE), v = runif(n = n, 0, 10000))
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
}
